% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FBAS.R
\name{FBAS}
\alias{FBAS}
\title{FBAS}
\usage{
FBAS(x, y, m, lambda, rho)
}
\arguments{
\item{x}{Exposure}

\item{y}{Outcome}

\item{m}{mediator}

\item{lambda}{regularizer}

\item{rho}{lagrangian penalty}
}
\value{

}
\description{
FBAS
}
